requirements.txt
absl-py==0.8.1
astor==0.8.0
attrs==19.3.0
audioread==2.1.8
backcall==0.1.0
bleach==3.1.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.13.0
click==7.1.1
colorama==0.4.3
cssselect2==0.2.2
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.1.1
EasyProcess==0.2.10
editdistance==0.5.3
entrypoint2==0.2
entrypoints==0.3
Flask==1.1.1
gast==0.3.2
google-images-download==2.8.0
grpcio==1.24.3
h5py==2.10.0
html5lib==1.0.1
imageio==2.8.0
importlib-metadata==1.5.0
inflect==2.1.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.0
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Lasagne==0.1
librosa==0.7.1
livelossplot==0.4.1
llvmlite==0.30.0
Mako==1.1.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
mock==3.0.5
MouseInfo==0.1.2
multiprocess==0.70.9
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.5
notebook==6.0.3
numba==0.46.0
numpy==1.16.5
olefile==0.46
opencv-python==4.2.0.32
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
Pillow==6.2.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.10.0
PyAutoGUI==0.9.48
pycparser==2.19
PyGetWindow==0.0.8
Pygments==2.5.2
pygpu==0.7.6
PyMsgBox==1.0.7
pyparsing==2.4.2
pyperclip==1.7.0
Pyphen==0.9.5
PyQt5==5.13.1
PyQt5-sip==12.7.0
PyRect==0.1.4
pyrsistent==0.15.7
pyscreenshot==1.0
PyScreeze==0.1.26
python-dateutil==2.8.0
PyTweening==1.0.3
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.1.2
pyzmq==19.0.0
resampy==0.2.2
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
sounddevice==0.3.14
SoundFile==0.10.2
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tflearn==0.3.2
Theano==0.7.0
tinycss2==1.0.2
torch==1.0.0
torchvision==0.2.1
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
umap-learn==0.3.10
Unidecode==1.1.1
urllib3==1.25.8
wcwidth==0.1.8
WeasyPrint==51
webencodings==0.5.1
webrtcvad==2.0.10
Werkzeug==0.16.0
wincertstore==0.2
zipp==3.1.0
